split ProjectedGradientDescent into a separate class
